requirements.txt
Flask==3.0.0
flask-cors
openai==1.14.3
twilio==9.5.2
python-dotenv==1.1.0
requests==2.32.3
schedule==1.2.2
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
gunicorn==20.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
gspread==6.0.0
oauth2client==4.1.3